% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataTrain.R
\name{dataTrain}
\alias{dataTrain}
\title{Data Generation}
\usage{
dataTrain(n, p, negativeProportion = 0)
}
\arguments{
\item{n}{the total number of rows in the data.}

\item{p}{the total number of columns in the data}

\item{negativeProportion}{the proportion of 'negative' examples in the data. Defaults to 0.}
}
\description{
This function simulates some data containing postive and negative examples.
}
\examples{
dataTrain(10, 5, 0.2)
}

