% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussianKern.R
\name{gaussianKern}
\alias{gaussianKern}
\title{Gaussian Kernel}
\usage{
gaussianKern(x, y, sigma)
}
\arguments{
\item{x}{a single observation (vector).}

\item{y}{a single observation (vector).}

\item{sigma}{the scaling parameter for the gaussian kernel}
}
\description{
This function calculates the gaussian kernel
}
\examples{
gaussianKern(c(2, 5), c(3, 2), sigma=1.5)
}

