% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gram_mat.R
\name{gram_mat}
\alias{gram_mat}
\title{Gram Matrix}
\usage{
gram_mat(mydat, sigma)
}
\arguments{
\item{mydat}{is the training data.}

\item{sigma}{is the scale parameter for use in the gaussian kernel.}
}
\description{
This function calculates the NxN gram matrix -- Kernel(X, X)
}
\examples{
trainingData <- dataTrain(10, 5)
gram_mat(trainingData, 1.5)
}

