% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernelDistance.R
\name{kernelDistance}
\alias{kernelDistance}
\title{Calculate the kernel distance}
\usage{
kernelDistance(point, data, alphas, gramMat, sigma)
}
\arguments{
\item{point}{is the observation we are calculating the distance for.}

\item{data}{is training data used to train the SVDD.}

\item{alphas}{is the vector of alphas output from the SVDD training function.}

\item{gramMat}{is the gram matrix from the training data.}

\item{sigma}{is the scale parameter for the gaussian kernel.}
}
\description{
This function calculates the kernel distance between an observation and the center of the SVDD.
}
\examples{
trainingData <- dataTrain(10, 5)
sig <- 1.5
gm <- gram_mat(trainingData, sig)
svdd <- svddTrain(trainingData, gm, sig, 1)
alphaValues <- svdd$alpha
kernelDistance(trainingData[1,], trainingData, alphaValues, gm, sig)
}

