% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makedvec.R
\name{make.d.vec}
\alias{make.d.vec}
\title{d vector}
\usage{
make.d.vec(mydat, sigma)
}
\arguments{
\item{mydat}{is the training data.}

\item{sigma}{is the scale parameter for use in the gaussian kernel.}
}
\description{
This is not to be called by the user. This function is for training the SVDD and is necessary for the quadratic programming problem.
}

