% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/svddTrain.R
\name{svddTrain}
\alias{svddTrain}
\title{Train the Support Vector Data Descriptor (SVDD)}
\usage{
svddTrain(X, Gram_Matrix, sigma, C1, C2 = 0, negativeProportion = 0)
}
\arguments{
\item{X}{is the training data.}

\item{Gram_Matrix}{is the gram matrix created by the gram_mat function.}

\item{sigma}{is the scale parameter for the gaussian kernel.}

\item{C1}{is the cost parameter for the postive examples.}

\item{C2}{is the cost parameter for the negative examples.}

\item{negativeProportion}{proportion of negative examples present in the training data.}
}
\description{
This function trains the SVDD model.
}
\examples{
trainingData <- dataTrain(10, 5)
sig <- 1.5
gm <- gram_mat(trainingData, sig)
svddTrain(trainingData, gm, sig, 1)
}

